Force DB connection attempt at activation time
The previous check, unless the database was already connected for
some reason, returned false and prevented custom roles to be loaded
at startup time.

So, before checking if the connection is available, the fix tries
to connect to the DB, and simply rescue any error that may arise,
as it happens when compiling assets on Docker. We can't really be
specific on the error even if we wanted to, as specific DB gems
raise different errors and we'd need to include all of them as
dependencies before being able to rescue each error class.

The idea is taken from https://github.com/globalize/globalize/pull/799
so all credits go to the author @djpremier.