Stop connecting to DB during `asset:precompile`
Because of how Rails engines are mounted when Rake tasks are being run,
the `Engine.load_custom_permissions` here is hit before the Rails server
boots--during any other deploy step.

This can make the `assets:precompile` Rake task complete *much faster*
in the case that the database cannot be connnected to and the
`ActiveRecord::Connection` times out after 60 seconds, for example.

---

More info about why I'm making this change, for the curious:

This makes any Rake task run as part of the application build process
depedent on the database when it doesn't need to be. This can cause
issues if the database is unavailable during the build step.

For example: the Heroku Ruby buildpack runs `assets:precompile` during
its build process. But Heroku build dynos do not have static or stable
IP addresses, meaning that the database may reject a connection or the
connection may time out.

Although `Engine.load_custom_permissions` already rescues from database
connection errors, this can be slow in the case that the connection
error is a time out.